LayoutTests/http/wpt/workers/promise-unhandled-rejection.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Reject

Harness Error (FAIL), message = Unhandled rejection

PASS UnhandledRejection event occurs if a rejected promise is not handled.
